Initial refactor of servers complete and need to test
