+++ server_python/asset_generation/images_leonardo.py
