+++ server_python/asset_generation/leonard_anime_styles.json
