+++ server_python/package.json
@@ -1,0 +1,21 @@
+{
+  "dependencies": {
+    "@types/express": "^5.0.0",
+    "axios": "^1.x.x",
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.0",
+    "express": "^4.21.2"
+  },
+  "scripts": {
+    "dev": "ts-node server.ts",
+    "start": "node dist/server.js",
+    "build": "tsc",
+    "watch": "ts-node-dev --respawn server.ts"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.17",
+    "@types/node": "^18.x.x",
+    "ts-node": "^10.x.x",
+    "typescript": "^4.x.x"
+  }
+}
