+++ server_python/prompt_chaining/step_2_create_content.py
