+++ server_python/prompt_chaining/step_4_create_profile_images.py
