+++ server_python/prompt_chaining/step_6_create_image_posts.py
