+++ server_python/templates/master.json
