+++ server_python/templates/profile_image_options.json
