+++ server_python/templates/profile_image_traits.json
