+++ server_python/templates/season.json
